# flex/coordinator/gs_flex_coordinator/openapi/openapi.yaml
openapi: 3.0.3
info:
  contact:
    email: graphscope@alibaba-inc.com
    name: GraphScope
  description: |-
    This is a specification for GraphScope FLEX HTTP service based on the OpenAPI 3.0 specification. You can find out more details about
    specification at [doc](https://swagger.io/specification/v3/).

    Some useful links:
    - [GraphScope Repository](https://github.com/alibaba/GraphScope)
    - [The Source API definition for GraphScope Interactive](https://github.com/GraphScope/portal/tree/main/httpservice)
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: GraphScope FLEX HTTP SERVICE API
  version: 0.9.1
...
servers:
- url: /
paths:
...
  /api/v1/graph/{graph_name}/datasource:
    get:
      description: List data source on graph
      operationId: get_datasource
      parameters:
      - explode: false
        in: path
        name: graph_name
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              example:
                vertices_datasource:
                - data_source: FILE
                  type_name: person
                  location: /home/graphscope/path/to/person.csv
                  property_mapping:
                    "0": id
                    "1": name
                - data_source: FILE
                  type_name: software
                  location: /home/graphscope/path/to/software.csv
                  property_mapping:
                    "0": id
                    "1": name
                edges_datasource:
                - data_source: FILE
                  type_name: knows
                  source_vertex: person
                  destination_vertex: person
                  location: /home/graphscope/path/to/person_knows_person.csv
                  source_pk_column_map:
                    "0": id
                  destination_pk_column_map:
                    "1": id
                  property_mapping:
                    "2": weight
                - data_source: FILE
                  type_name: created
                  source_vertex: person
                  destination_vertex: software
                  location: /home/graphscope/path/to/person_created_software.csv
                  source_pk_column_map:
                    "0": id
                  destination_pk_column_map:
                    "1": id
                  property_mapping:
                    "2": weight
              schema:
                $ref: '#/components/schemas/DataSource'
          description: Successful operation
      tags:
      - datasource
      x-openapi-router-controller: gs_flex_coordinator.controllers.datasource_controller
    post:
      description: Import data source in batch
      operationId: import_datasource
      parameters:
      - explode: false
        in: path
        name: graph_name
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DataSource'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
          description: successful operation
      tags:
      - datasource
      x-openapi-router-controller: gs_flex_coordinator.controllers.datasource_controller
  /api/v1/graph/{graph_name}/datasource/edge_datasource:
    post:
      description: Bind data source on edge type
      operationId: bind_edge_datasource
      parameters:
      - explode: false
        in: path
        name: graph_name
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EdgeDataSource'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
          description: successful operation
      tags:
      - datasource
      x-openapi-router-controller: gs_flex_coordinator.controllers.datasource_controller
  /api/v1/graph/{graph_name}/datasource/edge_datasource/{type_name}:
    delete:
      description: Unbind datasource on an edge type
      operationId: unbind_edge_datasource
      parameters:
      - explode: false
        in: path
        name: graph_name
        required: true
        schema:
          type: string
        style: simple
      - explode: false
        in: path
        name: type_name
        required: true
        schema:
          type: string
        style: simple
      - explode: true
        in: query
        name: source_vertex_type
        required: true
        schema:
          type: string
        style: form
      - explode: true
        in: query
        name: destination_vertex_type
        required: true
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
          description: Successful operation
      tags:
      - datasource
      x-openapi-router-controller: gs_flex_coordinator.controllers.datasource_controller
    get:
      description: Get edge data source
      operationId: get_edge_datasource
      parameters:
      - explode: false
        in: path
        name: graph_name
        required: true
        schema:
          type: string
        style: simple
      - explode: false
        in: path
        name: type_name
        required: true
        schema:
          type: string
        style: simple
      - explode: true
        in: query
        name: source_vertex_type
        required: true
        schema:
          type: string
        style: form
      - explode: true
        in: query
        name: destination_vertex_type
        required: true
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              example:
                data_source: FILE
                type_name: created
                source_vertex: person
                destination_vertex: software
                location: /home/graphscope/path/to/person_created_software.csv
                source_pk_column_map:
                  "0": id
                destination_pk_column_map:
                  "1": id
                property_mapping:
                  "2": weight
              schema:
                $ref: '#/components/schemas/EdgeDataSource'
          description: Successful operation
      tags:
      - datasource
      x-openapi-router-controller: gs_flex_coordinator.controllers.datasource_controller
  /api/v1/graph/{graph_name}/datasource/vertex_datasource:
    post:
      description: Bind data source on vertex type
      operationId: bind_vertex_datasource
      parameters:
      - explode: false
        in: path
        name: graph_name
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VertexDataSource'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
          description: successful operation
      tags:
      - datasource
      x-openapi-router-controller: gs_flex_coordinator.controllers.datasource_controller
  /api/v1/graph/{graph_name}/datasource/vertex_datasource/{type_name}:
    delete:
      description: Unbind datasource on a vertex type
      operationId: unbind_vertex_datasource
      parameters:
      - explode: false
        in: path
        name: graph_name
        required: true
        schema:
          type: string
        style: simple
      - explode: false
        in: path
        name: type_name
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
          description: Successful operation
      tags:
      - datasource
      x-openapi-router-controller: gs_flex_coordinator.controllers.datasource_controller
    get:
      description: Get vertex data source
      operationId: get_vertex_datasource
      parameters:
      - explode: false
        in: path
        name: graph_name
        required: true
        schema:
          type: string
        style: simple
      - explode: false
        in: path
        name: type_name
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              example:
                data_source: FILE
                type_name: person
                location: /home/graphscope/path/to/person.csv
                property_mapping:
                  "0": id
                  "1": name
              schema:
                $ref: '#/components/schemas/VertexDataSource'
          description: Successful operation
      tags:
      - datasource
      x-openapi-router-controller: gs_flex_coordinator.controllers.datasource_controller
...
components:
  schemas:
    ApiResponse:
      example: operation successful
      type: string
...
    VertexDataSource:
      example:
        type_name: type_name
        property_mapping:
          key: ""
        location: location
        data_source: ODPS
      properties:
        data_source:
          enum:
          - ODPS
          - FILE
          title: data_source
          type: string
        type_name:
          title: type_name
          type: string
        location:
          title: location
          type: string
        property_mapping:
          additionalProperties: true
          title: property_mapping
          type: object
      title: VertexDataSource
      type: object
      x-body-name: vertex_data_source
    EdgeDataSource:
      example:
        source_pk_column_map:
          key: ""
        type_name: type_name
        destination_pk_column_map:
          key: ""
        source_vertex: source_vertex
        destination_vertex: destination_vertex
        property_mapping:
          key: ""
        location: location
        data_source: ODPS
      properties:
        data_source:
          enum:
          - ODPS
          - FILE
          title: data_source
          type: string
        type_name:
          title: type_name
          type: string
        source_vertex:
          title: source_vertex
          type: string
        destination_vertex:
          title: destination_vertex
          type: string
        location:
          title: location
          type: string
        source_pk_column_map:
          additionalProperties: true
          title: source_pk_column_map
          type: object
        destination_pk_column_map:
          additionalProperties: true
          title: destination_pk_column_map
          type: object
        property_mapping:
          additionalProperties: true
          title: property_mapping
          type: object
      title: EdgeDataSource
      type: object
      x-body-name: edge_data_source
    DataSource:
      example:
        edges_datasource:
        - source_pk_column_map:
            key: ""
          type_name: type_name
          destination_pk_column_map:
            key: ""
          source_vertex: source_vertex
          destination_vertex: destination_vertex
          property_mapping:
            key: ""
          location: location
          data_source: ODPS
        - source_pk_column_map:
            key: ""
          type_name: type_name
          destination_pk_column_map:
            key: ""
          source_vertex: source_vertex
          destination_vertex: destination_vertex
          property_mapping:
            key: ""
          location: location
          data_source: ODPS
        vertices_datasource:
        - type_name: type_name
          property_mapping:
            key: ""
          location: location
          data_source: ODPS
        - type_name: type_name
          property_mapping:
            key: ""
          location: location
          data_source: ODPS
      properties:
        vertices_datasource:
          items:
            $ref: '#/components/schemas/VertexDataSource'
          title: vertices_datasource
          type: array
        edges_datasource:
          items:
            $ref: '#/components/schemas/EdgeDataSource'
          title: edges_datasource
          type: array
      title: DataSource
      type: object
      x-body-name: data_source
...